Add support for Scala 2.11/2.12 in worksheets
Previously, worksheets only worked for Scala 2.12.10. Now, worksheets
support all Scala versions where completions are supported.

Another nice improvements in this change is that we no longer poke into
internal mdoc APIs. The worksheet evaluation logic has been added to
public mdoc APIs so that we have a clear separation of concerns. This
means that improvements to rendering of values in worksheets need to be
contributed to the scalameta/mdoc repository instead of
scalameta/metals.
