+++ metals/src/main/scala/scala/meta/internal/metals/MetalsLanguageServer.scala
@@ -417,7 +417,8 @@ class MetalsLanguageServer(
             languageClient,
             () => userConfig,
             statusBar,
-            diagnostics
+            diagnostics,
+            embedded
           )
         )
       )
