+++ metals/src/main/scala/scala/meta/internal/worksheets/WorksheetClassLoader.scala
@@ -1,0 +1,15 @@
+package scala.meta.internal.worksheets
+
+/**
+ * ClassLoader that is used to reflectively invoke mdoc for worksheet evaluation.
+ */
+class MdocClassLoader(parent: ClassLoader) extends ClassLoader(null) {
+  override def findClass(name: String): Class[_] = {
+    val isShared = name.startsWith("mdoc.interfaces")
+    if (isShared) {
+      parent.loadClass(name)
+    } else {
+      throw new ClassNotFoundException(name)
+    }
+  }
+}
