+++ metals/src/main/scala/scala/meta/internal/worksheets/WorksheetProvider.scala
@@ -9,22 +9,12 @@ import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.Cancelable
 import scala.collection.concurrent.TrieMap
 import ch.epfl.scala.bsp4j.BuildTargetIdentifier
-import mdoc.internal.cli.Context
-import mdoc.internal.markdown.SectionInput
-import mdoc.internal.markdown.Modifier
-import mdoc.internal.markdown.Instrumenter
-import mdoc.internal.markdown.MarkdownCompiler
 import org.eclipse.{lsp4j => l}
 import scala.meta.internal.decorations.ThemableDecorationInstanceRenderOptions
 import scala.meta.internal.decorations.ThemableDecorationAttachmentRenderOptions
-import mdoc.internal.cli.Settings
-import scala.meta.internal.metals.ScalaTarget
 import scala.concurrent.Future
 import scala.meta.pc.CancelToken
-import pprint.TPrintColors
 import scala.meta.internal.metals.UserConfiguration
-import scala.collection.immutable.Nil
-import mdoc.document.Statement
 import scala.meta.internal.metals.MetalsLanguageClient
 import scala.meta.internal.metals.ScalaVersions
 import scala.meta.internal.metals.BuildInfo
@@ -40,9 +30,11 @@ import scala.meta.internal.pc.InterruptException
 import scala.util.control.NonFatal
 import java.util.concurrent.Executors
 import scala.meta.internal.metals.Diagnostics
-import pprint.PPrinter.BlackWhite
 import scala.meta.internal.metals.Timer
 import scala.meta.internal.metals.Time
+import scala.meta.internal.metals.Embedded
+import mdoc.{interfaces => i}
+import mdoc.interfaces.Mdoc
 
 /**
  * Implements interactive worksheets for "*.worksheet.sc" file extensions.
@@ -56,14 +48,11 @@ class WorksheetProvider(
     languageClient: MetalsLanguageClient,
     userConfig: () => UserConfiguration,
     statusBar: StatusBar,
-    diagnostics: Diagnostics
+    diagnostics: Diagnostics,
+    embedded: Embedded
 )(implicit ec: ExecutionContext)
     extends Cancelable {
 
-  private val commentHeader = " // "
-  // The smallest column width that worksheet values will use for rendering
-  // worksheet decorations.
-  private val minimumMargin = 20
   // Worksheet evaluation happens on a single threaded job queue. Jobs are
   // prioritized using the same order as completion/hover requests:
   // first-come last-out.
@@ -74,19 +63,18 @@ class WorksheetProvider(
   private lazy val threadStopper: ScheduledExecutorService =
     Executors.newSingleThreadScheduledExecutor()
   private val cancelables = new MutableCancelable()
-  private val contexts = new TrieMap[BuildTargetIdentifier, Context]()
-  private val reporter = new MdocStoreReporter()
+  private val mdocs = new TrieMap[BuildTargetIdentifier, Mdoc]()
 
   def onBuildTargetDidCompile(target: BuildTargetIdentifier): Unit = {
     clearBuildTarget(target)
   }
 
   private def clearBuildTarget(target: BuildTargetIdentifier): Unit = {
-    contexts.remove(target).foreach(_.compiler.global.close())
+    mdocs.remove(target).foreach(_.shutdown())
   }
 
   def reset(): Unit = {
-    contexts.keysIterator.foreach(clearBuildTarget)
+    mdocs.keysIterator.foreach(clearBuildTarget)
   }
 
   def cancel(): Unit = {
@@ -99,7 +87,6 @@ class WorksheetProvider(
       path: AbsolutePath,
       token: CancelToken
   ): Future[Array[DecorationOptions]] = {
-    reporter.reset()
     val result = new CompletableFuture[Array[DecorationOptions]]()
     def completeEmptyResult() = result.complete(Array.empty)
     token.onCancel().asScala.foreach { isCancelled =>
@@ -216,179 +203,103 @@ class WorksheetProvider(
     )
   }
 
-  private def renderHoverMessage(
-      statement: Statement,
-      margin: Int,
-      isEmptyValue: Boolean
-  ): String = {
-    val out = new StringBuilder()
-    if (!isEmptyValue) {
-      statement.binders.iterator.foreach { binder =>
-        out
-          .append("\n")
-          .append(binder.name)
-          .append(": ")
-          .append(binder.tpe.render(TPrintColors.BlackWhite))
-          .append(" = ")
-        BlackWhite
-          .tokenize(binder.value, width = 100)
-          .foreach(text => out.appendAll(text.getChars))
-      }
-    }
-    statement.out.linesIterator.foreach { line =>
-      out.append("\n// ").append(line)
-    }
-    out.toString()
-  }
-
-  private def renderContentText(
-      statement: Statement,
-      margin: Int,
-      isEmptyValue: Boolean
-  ): String = {
-    val out = new StringBuilder()
-    out.append(commentHeader)
-    if (isEmptyValue) {
-      if (!statement.out.isEmpty()) {
-        out.append(statement.out.linesIterator.next())
-      }
-    } else {
-      val isSingle = statement.binders.lengthCompare(1) == 0
-      statement.binders.iterator.zipWithIndex.foreach {
-        case (binder, i) =>
-          if (!isSingle) {
-            out
-              .append(if (i == 0) "" else ", ")
-              .append(binder.name)
-              .append("=")
-          }
-          val truncatedLine = BlackWhite
-            .tokenize(binder.value, width = margin)
-            .map(_.getChars)
-            .filterNot(_.iterator.forall(_.isWhitespace))
-            .flatMap(_.iterator)
-            .filter {
-              case '\n' => false
-              case _ => true
-            }
-            .take(margin)
-          out.appendAll(truncatedLine)
-      }
-    }
-    out.toString()
-  }
-
   private def evaluateWorksheet(
       path: AbsolutePath,
       token: CancelToken
   ): Array[DecorationOptions] = {
-    val input = path.toInputFromBuffers(buffers)
-    // NOTE(olafurpg): the sbt dialects is the closest available syntax to worksheets.
-    val WorksheetDialect = dialects.Sbt1
     val decorations = for {
-      ctx <- getContext(path)
-      source <- WorksheetDialect(input).parse[Source].toOption
-    } yield renderDecorations(path, ctx, source)
-    decorations.getOrElse(Array.empty)
-  }
-
-  private def renderDecorations(
-      path: AbsolutePath,
-      ctx: Context,
-      source: Source
-  ): Array[DecorationOptions] = {
-    val sectionInput = SectionInput(
-      path.toInputFromBuffers(buffers),
-      source,
-      Modifier.Default()
-    )
-    val sectionInputs = List(sectionInput)
-    val instrumented = Instrumenter.instrument(sectionInputs)
-    val rendered = MarkdownCompiler.buildDocument(
-      ctx.compiler,
-      ctx.reporter,
-      sectionInputs,
-      instrumented,
-      path.toString
-    )
-
-    val decorations = for {
-      section <- rendered.sections.iterator
-      statement <- section.section.statements
-    } yield renderDecoration(statement)
-
-    diagnostics.onPublishDiagnostics(
-      path,
-      reporter.diagnostics.toSeq,
-      isReset = true
-    )
-
-    decorations
-      .filterNot(_.renderOptions.after.contentText == commentHeader)
-      .toArray
-  }
-
-  private def renderDecoration(statement: Statement): DecorationOptions = {
-    val pos = statement.position
-    val range = new l.Range(
-      new l.Position(pos.startLine, pos.startColumn),
-      new l.Position(pos.endLine, pos.endColumn)
-    )
-    val margin = math.max(
-      minimumMargin,
-      userConfig().worksheetScreenWidth - statement.position.endColumn
-    )
-    val isEmptyValue = isUnitType(statement) || statement.binders.isEmpty
-    val contentText = renderContentText(statement, margin, isEmptyValue)
-    val hoverMessage = renderHoverMessage(statement, margin, isEmptyValue)
-    DecorationOptions(
-      range,
-      new l.MarkedString("scala", hoverMessage),
-      ThemableDecorationInstanceRenderOptions(
-        after = ThemableDecorationAttachmentRenderOptions(
-          contentText,
-          color = "green",
-          fontStyle = "italic"
+      mdoc <- getMdoc(path)
+      input = path.toInputFromBuffers(buffers)
+    } yield mdoc.evaluateWorksheet(input.path, input.value)
+    decorations match {
+      case None => Array.empty
+      case Some(worksheet) =>
+        diagnostics.onPublishDiagnostics(
+          path,
+          worksheet.diagnostics().iterator().asScala.map(toLsp).toSeq,
+          isReset = true
         )
-      )
-    )
+        worksheet
+          .statements()
+          .iterator()
+          .asScala
+          .map { s =>
+            new DecorationOptions(
+              toLsp(s.position()),
+              new l.MarkedString("scala", s.details()),
+              ThemableDecorationInstanceRenderOptions(
+                after = ThemableDecorationAttachmentRenderOptions(
+                  s.summary(),
+                  color = "green",
+                  fontStyle = "italic"
+                )
+              )
+            )
+          }
+          .toArray
+    }
   }
 
-  private def getContext(path: AbsolutePath): Option[Context] = {
+  private def toLsp(p: i.RangePosition): l.Range = {
+    new l.Range(
+      new l.Position(
+        p.startLine(),
+        p.startColumn()
+      ),
+      new l.Position(
+        p.endLine(),
+        p.endColumn()
+      )
+    ),
+  }
+  private def toLsp(d: i.Diagnostic): l.Diagnostic = {
+    new l.Diagnostic(
+      toLsp(d.position()),
+      d.message(),
+      d.severity() match {
+        case i.DiagnosticSeverity.Info => l.DiagnosticSeverity.Information
+        case i.DiagnosticSeverity.Warning => l.DiagnosticSeverity.Warning
+        case i.DiagnosticSeverity.Error => l.DiagnosticSeverity.Error
+        case _ => l.DiagnosticSeverity.Error
+      },
+      "mdoc"
+    )
+
+  }
+
+  private def getMdoc(path: AbsolutePath): Option[Mdoc] = {
     for {
       target <- buildTargets.inverseSources(path)
-      info <- buildTargets.scalaTarget(target)
-      scala <- info.info.asScalaBuildTarget
-      scalaVersion = scala.getScalaVersion
-      isSupported = ScalaVersions.isCurrentScalaCompilerVersion(scalaVersion)
-      _ = {
-        if (!isSupported) {
-          scribe.warn(
-            s"worksheet: unsupported Scala version '${scalaVersion}', to fix this problem use Scala version '${BuildInfo.scala212}' instead."
-          )
+      mdoc <- getMdoc(target)
+    } yield mdoc
+  }
+  private def getMdoc(target: BuildTargetIdentifier): Option[Mdoc] = {
+    mdocs.get(target).orElse {
+      for {
+        info <- buildTargets.scalaTarget(target)
+        scala <- info.info.asScalaBuildTarget
+        scalaVersion = scala.getScalaVersion
+        isSupported = ScalaVersions.isSupportedScalaVersion(scalaVersion)
+        _ = {
+          if (!isSupported) {
+            scribe.warn(
+              s"worksheet: unsupported Scala version '${scalaVersion}', to fix this problem use Scala version '${BuildInfo.scala212}' instead."
+            )
+          }
         }
+        if isSupported
+      } yield {
+        val scalacOptions = info.scalac.getOptions.asScala
+          .filterNot(_.contains("semanticdb"))
+          .asJava
+        val mdoc = embedded
+          .mdoc(scala)
+          .withClasspath(info.fullClasspath.asScala.distinct.asJava)
+          .withScalacOptions(scalacOptions)
+        mdocs(target) = mdoc
+        mdoc
       }
-      if isSupported
-    } yield contexts.getOrElseUpdate(target, newContext(target, info))
-  }
-
-  private def newContext(
-      target: BuildTargetIdentifier,
-      info: ScalaTarget
-  ): Context = {
-    scribe.info(s"worksheet: new compiler for ${target.getUri()}")
-    val settings = Settings.default(workspace)
-    val classpath = Classpath(info.fullClasspath).syntax
-    val options = info.scalac.getOptions().asScala.mkString(" ")
-    val compiler = MarkdownCompiler.fromClasspath(classpath, options)
-    Context(settings, reporter, compiler)
-  }
-
-  private def isUnitType(statement: Statement): Boolean = {
-    statement.binders match {
-      case head :: Nil => head.value == ()
-      case _ => false
     }
-
   }
+
 }
