+++ tests/slow/src/test/scala/tests/feature/WorksheetCrossLspSuite.scala
@@ -1,0 +1,6 @@
+package tests.feature
+
+import scala.meta.internal.metals.{BuildInfo => V}
+
+object Worksheet211LspSuite extends tests.BaseWorksheetLspSuite(V.scala211)
+object Worksheet213LspSuite extends tests.BaseWorksheetLspSuite(V.scala213)
